+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * `/v3/limits` endpoint to get limits list with filters
 * Squash matches with the same price in history
 * `/v3/movements` endpoint to get participants effect related to balance movements
+* `/v3` endpoint to get horizon info
 
 ## Fixed
 * `/v3/account` to include all limits imposed for account